requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-multipart
python-dotenv
transformers
torch
pillow
opencv-python
tesseract
pytesseract
alembic
